requirements.txt
aniso8601==7.0.0
Authlib==1.3.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
flask==3.0.3
Flask-Cors==4.0.1
Flask-GraphQL==2.0.1
Flask-PyMongo==2.3.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==3.7
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
promise==2.3
pycparser==2.22
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
Rx==1.6.3
six==1.16.0
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
